+++ atlasdb-cassandra-integration-tests/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/CassandraKeyValueServiceSerializableTransactionIntegrationTest.java
@@ -26,6 +26,7 @@ import com.palantir.atlasdb.containers.Containers;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.sweep.queue.MultiTableSweepQueueWriter;
 import com.palantir.atlasdb.sweep.queue.SpecialTimestampsSupplier;
+import com.palantir.atlasdb.sweep.queue.TargetedSweepFollower;
 import com.palantir.atlasdb.sweep.queue.TargetedSweeper;
 import com.palantir.atlasdb.transaction.impl.AbstractSerializableTransactionTest;
 
@@ -51,7 +52,8 @@ public class CassandraKeyValueServiceSerializableTransactionIntegrationTest
     @Override
     protected MultiTableSweepQueueWriter getSweepQueueWriterInitialized() {
         TargetedSweeper queue = TargetedSweeper.createUninitialized(() -> true, () -> 128, 0, 0, mock(Follower.class));
-        queue.initialize(new SpecialTimestampsSupplier(() -> 0, () -> 0), keyValueService, null);
+        queue.initialize(new SpecialTimestampsSupplier(() -> 0, () -> 0), keyValueService,
+                mock(TargetedSweepFollower.class));
         return queue;
     }