+++ atlasdb-tests-shared/src/test/java/com/palantir/atlasdb/AtlasDbTestCase.java
@@ -81,7 +81,6 @@ public class AtlasDbTestCase {
     protected TargetedSweeper sweepQueue;
     protected int sweepQueueShards = 128;
     protected ExecutorService deleteExecutor = Executors.newSingleThreadExecutor();
-    protected Follower follower = mock(Follower.class);
 
     @BeforeClass
     public static void setupLockClient() {
@@ -118,7 +117,8 @@ public class AtlasDbTestCase {
         conflictDetectionManager = ConflictDetectionManagers.createWithoutWarmingCache(keyValueService);
         sweepStrategyManager = SweepStrategyManagers.createDefault(keyValueService);
 
-        sweepQueue = spy(TargetedSweeper.createUninitialized(() -> true, () -> sweepQueueShards, 0, 0, follower));
+        sweepQueue = spy(
+                TargetedSweeper.createUninitialized(() -> true, () -> sweepQueueShards, 0, 0, mock(Follower.class)));
 
         serializableTxManager = new TestTransactionManagerImpl(
                 keyValueService,