+++ atlasdb-tests-shared/src/test/java/com/palantir/atlasdb/schema/stream/StreamTest.java
@@ -62,7 +62,6 @@ import com.google.common.collect.Sets;
 import com.google.common.util.concurrent.Futures;
 import com.google.protobuf.ByteString;
 import com.palantir.atlasdb.AtlasDbTestCase;
-import com.palantir.atlasdb.cleaner.CleanupFollower;
 import com.palantir.atlasdb.encoding.PtBytes;
 import com.palantir.atlasdb.protos.generated.StreamPersistence;
 import com.palantir.atlasdb.protos.generated.StreamPersistence.StreamMetadata;
@@ -86,7 +85,6 @@ import com.palantir.atlasdb.schema.stream.generated.TestHashComponentsStreamMeta
 import com.palantir.atlasdb.schema.stream.generated.TestHashComponentsStreamStore;
 import com.palantir.atlasdb.schema.stream.generated.TestHashComponentsStreamValueTable.TestHashComponentsStreamValueRow;
 import com.palantir.atlasdb.stream.PersistentStreamStore;
-import com.palantir.atlasdb.sweep.queue.ShardAndStrategy;
 import com.palantir.atlasdb.table.description.Schemas;
 import com.palantir.atlasdb.transaction.api.Transaction;
 import com.palantir.atlasdb.transaction.api.TransactionConflictException;
@@ -118,11 +116,7 @@ public class StreamTest extends AtlasDbTestCase {
     public TemporaryFolder temporaryFolder = new TemporaryFolder();
 
     @Before
-    public void createSchema() throws Exception {
-        sweepQueueShards = 1;
-        follower = CleanupFollower.create(StreamTestSchema.getSchema());
-        super.setUp();
-        serializableTxManager.setUnreadableTimestamp(Long.MAX_VALUE);
+    public void createSchema() {
         Schemas.deleteTablesAndIndexes(StreamTestSchema.getSchema(), keyValueService);
         Schemas.createTablesAndIndexes(StreamTestSchema.getSchema(), keyValueService);
 
@@ -158,11 +152,6 @@ public class StreamTest extends AtlasDbTestCase {
             assertArrayEquals(data, outputStream.toByteArray());
             return null;
         });
-        txManager.runTaskWithRetry(transaction -> {
-            defaultStore.unmarkStreamAsUsed(transaction, streamId, "ref".getBytes());
-            return null;}
-        );
-        sweepQueue.sweepNextBatch(ShardAndStrategy.conservative(0));
     }
 
     @Test